Move kubeDetailItems impl to file
